Fix: Display up to 4 campsites per row.
Modify app.css to set the max width of the camp card container to 1500px to limit the number of campsites per row.

Add a .camp-list selector, and set the display to a grid, with centered content so the search bar continues to display at the top of the page.